// src/models/relatorio.ts
export class Relatorio {
  id: number;
  fantasia: string;
  razao: string;
  cnpj: string;
  endereco: string;
  cep: string;
  bairro: string;
  municipio: string;
  uf: string;
  telefone: string;
  fax: string;
  obs: string;
  inpesao: string;
  data_inspecao: string;
  data_relatorio: string;
  objetivo: string;
  providencias_tomadas: string;
  pessoas_constatadas: string;
  relato_situacao: string;
  irregularidades: any;
  orientacoes_gerais: string;
  consideracoes: string;
  tipo_inspecao: string;
  situacao: string;
  fiscal_responsavel: string;
  fiscais_presentes: [];
}